+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 Version 1.6.1 ()
 ===========================================================
 
+*   Feature: Can now try and auto-detect local torrent clients
 *   Feature: Now possible to have multiple clients configured
 *   Bugfix: Fixed small bugs with Deluge client
 *   Bugfix: Dry run tried to connect to the client